pyproject.toml
[tool.poetry]
name = "camsensex1"
version = "0.1.0"
description = ""
authors = ["Takahiro55555"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
websocket-client = "^0.57.0"
pyserial = "^3.4"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
